ads1118: Add support for ADS1118 ADC chip
Added firmware and klippy driver for Texas Instruments ADS1118 ADC chip.
The ADS1118 is an ADC chip with an SPI interface. It features an
internal temperature sensor and can be used to read temperatures from
thermocouples. Implemented are reading from the internal
temperature sensor and k-type thermocouples.

The Flashforge Creator Pro2 printer uses this chip for measuring
extruder temperatures.

Signed-off-by: Christoph Nelles <klipper-fw@evilazrael.de>